move sync actions test to separate file
